// src/lostark/auctions/dto/auction-search-option.dto.ts
interface SearchDetailOption {
  FirstOption: number;
  SecondOption: number;
  MinValue?: number;
  MaxValue?: number;
}

export class AuctionSearchOptionDto {
  Sort: string;
  SortCondition: string;
  ItemTier: number;
  PageNo: number;
  CategoryCode: number;
  ItemGradeQuality: number;
  ItemGrade: string;
  ItemName: string;
  SkillOptions: SearchDetailOption[];
  EtcOptions: SearchDetailOption[];
}